Intermediate commit: Start changes to linear mixing and learn PyMC3
It would seem that PyMC3 has a function that does precisely what I want
for an arbitraty number of distributions.
This does give an indication that there are other areas in literature to
look and gain an intuition (and build a defense) for the case of
likelihood mixing.

The next step is to build out an Tawere-like interface for Global mixing
and figure out how to do the local mixing with the `pm.Mixure` model.